+++ packages/phone-number-privacy/combiner/test/integration/pnp.test.ts
@@ -76,20 +76,11 @@ combinerConfig.phoneNumberPrivacy.enabled = true
 const signerConfig: SignerConfig = {
   serviceName: 'odis-signer',
   server: {
-    port: undefined,
+    port: 8080, // actual port specified with .listen() in test setup
     sslKeyPath: undefined,
     sslCertPath: undefined,
   },
   quota: {
-    unverifiedQueryMax: 10,
-    additionalVerifiedQueryMax: 30,
-    queryPerTransaction: 2,
-    // Min balance is .01 cUSD
-    minDollarBalance: new BigNumber(1e16),
-    // Min balance is .01 cEUR
-    minEuroBalance: new BigNumber(1e16),
-    // Min balance is .005 CELO
-    minCeloBalance: new BigNumber(5e15),
     // Equivalent to 0.001 cUSD/query
     queryPriceInCUSD: new BigNumber(0.001),
   },